move Types into TypeRef
